% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/cd.R
\name{cd}
\alias{cd}
\title{Change Directory}
\usage{
cd(filepath)
}
\arguments{
\item{filepath}{Filepath to set as current working directory.}
}
\description{
Wrapper for setwd() using unix command form.
}
\keyword{wrapper}
