% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/doy_2md.R
\name{doy_2md}
\alias{doy_2md}
\title{Converting day of year number to month and day}
\usage{
doy_2md(doy, short = FALSE, leap = FALSE)
}
\arguments{
\item{doy}{numeric or vector of numerics representing the day of year.}

\item{short}{Do we want months abbreviated? Defaults to \code{FALSE}; if \code{TRUE}, longer month names like January will be represented with their abbreviations. Useful for axes labels.}

\item{leap}{Do we want to assume a leap year? Defaults to \code{FALSE}; if \code{TRUE}, uses a 29-day February.}
}
\value{
A character string or vector of strings, with human readable day/month.
}
\description{
Translates numbers into human-readable day and month, useful when working with phenology data.
Note that this function treats decimals as their floor, so \code{doy_2md(doy = 5.8)} returns "January 05".
By default, assumes a non-leap-year.
}
\examples{
doy_2md(18)
doy_2md(c(15, 29.8, 180))
doy_2md(c(15, 29.8, 180), short=TRUE)
}
\keyword{date}
