% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ggplot_lister.R
\name{ggplot_list_saver}
\alias{ggplot_list_saver}
\title{Create and save sequential (or branching) ggplot figures}
\usage{
ggplot_list_saver(
  gglist,
  base.name,
  filetype = ".jpg",
  save.path = ".",
  suffix = "-V",
  niceties = NULL,
  ...
)
}
\arguments{
\item{gglist}{A list of ggplot layers, augmented with whole numbers to specify commands. First item in the list must be a \code{ggplot()} call.
\code{0} specifies printing, negative integers specify removing the previous \code{n} layers before continuing. Each item MUST be either a \code{ggplot()} call, ggplot layer, \code{0}, or negative whole number.}

\item{base.name}{Name for plots (numbers will be added) without filetype.}

\item{filetype}{suffix for file type, including period. Default is ".jpg"}

\item{save.path}{Path to save files to. Default is current working directory.}

\item{suffix}{Optional suffix to separate base name from plot numbers. Default is "-V". Set to "" to ignore.}

\item{niceties}{A single ggplot layer or a list of ggplot layers that should be applied to all plots, and
are applied last. Common use case would be a list of axis labels, plot title, and theme. Defaults to NULL; in this case, no layers added.
Note that this formulation is merely for convenience -- these layers can also be specified in gglist, but would generally need to be added before the first \code{0} entry.}

\item{...}{Additional arguments passed to ggsave. Likely choices: width, height, units, dpi. See \code{?ggsave} for details.}
}
\value{
Logical
}
\description{
Create and save sequential (or branching) ggplot figures
}
\examples{

library(tidyverse)
dat = data.frame(y = runif(30), x = runif(30))
# ggplot_list_saver(gglist = list(ggplot(data = dat, aes(x = x, y = y)),
#   geom_point(), 0,
#   geom_path(),0, -1, geom_smooth(method = 'lm')),
#   niceties = list(xlab("turtle"), ylab("cow"), theme_bw()),
#   base.name = "testfig",
#   save.path = "."
# )
}
