% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/note_extractor.R
\name{note_extractor}
\alias{note_extractor}
\title{Extract notes from R scripts}
\usage{
note_extractor(infile, outfile = NULL, marker = "#\%note")
}
\arguments{
\item{infile}{Name of the file to search through for notes}

\item{outfile}{Name of the file notes will be saved to. Default is \code{NULL}, and in this case the outfile will be set to the infile name, but with an ending of \verb{'-notes.txt}}

\item{marker}{String used to identify comment lines that have notes. Defaults to \verb{#\%note}.}
}
\description{
Searches a text file (including R and Rmarkdown) for lines with a specific marker
(default, \code{'#note'}), and saves those lines to a new text file. Useful
for writing notes in a long analysis or simulation script. Example usecase:
my analysis script of 35+ species of butterflies -- writing notes for each species that has problematic fit diagnostics.
While this can be done in a separate file, leaving the notes in the script allows for natural updating.
note_extractor makes it easy to access these notes.
}
