% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/skeleton_maker.R
\name{skeleton_maker}
\alias{skeleton_maker}
\title{Project directory maker}
\usage{
skeleton_maker(project.name, filepath = ".")
}
\arguments{
\item{project.name}{Name of the project itself, used for the root directory of the new project.}

\item{filepath}{filepath for where the root directory will live. Defaults to working directory.}
}
\description{
This function creates my preferred directory structure and key files for new research projects
In addition to folders, this creates:
\itemize{
\item a blank datasources.txt file in the \verb{1_raw_data}, to keep track of where data files come from
\item a blank funs.R file in \verb{3_scripts}
\item a template file for recommended reviewers (\code{Reviewers.docx}) in \code{ms} -- this makes life easier when resubmitting
\item a template code map file (\verb{Code map.txt}) for mapping out what files go where, with reminder text on how to generate a directory tree.
}
}
\examples{
skeleton_maker("unicorn-pop-dynamics")
}
\keyword{organization}
